_Summarise the changes this Pull Request makes._

Re-use a socket. This is not a user-visible breaking change, so could be a patch version.
However we need to find out if the socket is thread-safe, if there are any cases where a new socket is needs and how to do that in a thread-safe manner, and how to dispose of the socket.

Why would we do this? Check the before and after benchmarks:

Before:

|             Method |      Mean |    Error |   StdDev |  Gen 0 | Allocated |
|------------------- |----------:|---------:|---------:|-------:|----------:|
|              RunIp | 372.77 us | 4.851 us | 3.788 us |      - |    1184 B |
|  RunIpWithSampling |  96.21 us | 1.765 us | 1.651 us | 0.1221 |     506 B |

After:

|             Method |       Mean |     Error |    StdDev |  Gen 0 | Allocated |
------------------- |-----------:|----------:|----------:|-------:|----------:|
|              RunIp |  16.352 us | 0.0783 us | 0.0611 us | 0.2136 |     720 B |
|  RunIpWithSampling |   5.224 us | 0.0745 us | 0.0697 us | 0.1221 |     388 B |