@jmahmud 2018-07-16
@AnthonySteele - does a mechanism for re-creating a socket be part of this implementation - catching of exception / on certain conditions?  Not sure what those would be however...