@AnthonySteele 2018-07-16
Yes, throwing the socket away if there's an error is possible, but we first need to know if `  socket.SendTo(bytes, endpoint)` is completely thread-safe and re-entrant, like `HttpClient.SendAsync`.