@martincostello 2018-07-16
Maybe worth asking the question in the corefx (or whatever) repo? Alternatively, maybe have a cursory dig through the new `HttpClient` stack that's built on top of sockets and see what they do with them.