@dv00d00 2018-07-16
Might be worth consideration:

|Method |     Mean |    Error |   StdDev | Scaled | ScaledSD | Allocated |
|------- |---------:|---------:|---------:|-------:|---------:|----------:|
| SendTo | 64.73 us | 179.1 us | 10.12 us |   1.00 |     0.00 |     168 B |
|   Send | 51.24 us | 255.5 us | 14.44 us |   0.80 |     0.21 |       0 B |


```csharp
    [MemoryDiagnoser]
    public class SocketTests
    {
        private Socket socket;
        private IPEndPoint ipEndPoint;
        private byte[] bytes;

        [GlobalSetup(Target = nameof(SendTo))]
        public void Setup()
        {
            socket = new Socket(AddressFamily.InterNetwork, SocketType.Dgram, ProtocolType.Udp);
            ipEndPoint = new IPEndPoint(new IPAddress(new byte[] { 127, 0, 0, 1 }), 8888);
            bytes = Encoding.ASCII.GetBytes("hello world yo");
        }

        [GlobalSetup(Target = nameof(Send))]
        public void SetupSend()
        {
            socket = new Socket(AddressFamily.InterNetwork, SocketType.Dgram, ProtocolType.Udp);
            ipEndPoint = new IPEndPoint(new IPAddress(new byte[] { 127, 0, 0, 1 }), 8888);
            bytes = Encoding.ASCII.GetBytes("hello world yo");
            socket.Connect(ipEndPoint);
        }

        [Benchmark(Baseline = true)]
        public void SendTo()
        {
            socket.SendTo(bytes, ipEndPoint);
        }

        [Benchmark]
        public void Send()
        {
            socket.Send(bytes);
        }
    }
```