@dv00d00 2018-07-16
Also found this in socket handler for http client

https://github.com/dotnet/corefx/blob/master/src/System.Net.Http/src/System/Net/Http/SocketsHttpHandler/ConnectHelper.cs

might be useful. But they are dealing with tcp sockets, which are stateful, UDP socket might be much simpler to deal with.