@AnthonySteele 2018-07-16
[Inquiry about sockets](https://github.com/dotnet/corefx/issues/29932)