@AnthonySteele 2018-07-18
[No, sockets are not thread-safe](https://github.com/dotnet/corefx/issues/31117). If we do anything like this, it should be via a pool.