@martincostello 2018-07-18
Here's the pooling implementation I've been using internally, based on the `2.1.0` codebase:

```csharp
using System;
using System.Collections.Generic;
using System.Linq;
using System.Net.Sockets;
using System.Runtime.InteropServices;
using JustEat.StatsD.EndpointLookups;

namespace JustEat.StatsD
{
    /// <summary>
    /// A class representing an implementation of <see cref="IStatsDTransport"/> that pools UDP clients. This class cannot be inherited.
    /// </summary>
    public sealed class PooledUdpTransport : IStatsDTransport, IDisposable
    {
        private const int PoolSize = 30;

        private static readonly SimpleObjectPool<SocketAsyncEventArgs> EventArgsPool
            = new SimpleObjectPool<SocketAsyncEventArgs>(PoolSize, pool => new PoolAwareSocketAsyncEventArgs(pool));

        private readonly SimpleObjectPool<UdpClient> _clientPool
            = new SimpleObjectPool<UdpClient>(PoolSize, pool => CreateUdpClient());

        private readonly IPEndPointSource _endpointSource;

        /// <summary>
        /// Initializes a new instance of the <see cref="PooledUdpTransport"/> class.
        /// </summary>
        /// <param name="endPointSource">The <see cref="IPEndPointSource"/> to use.</param>
        /// <exception cref="ArgumentNullException">
        /// <paramref name="endPointSource"/> is <see langword="null"/>.
        /// </exception>
        public PooledUdpTransport(IPEndPointSource endPointSource)
        {
            _endpointSource = endPointSource ?? throw new ArgumentNullException(nameof(endPointSource));
        }

        /// <inheritdoc />
        public void Dispose()
        {
            UdpClient client = null;

            while ((client = _clientPool.Pop()) != null)
            {
                client.Dispose();
            }
        }

        /// <inheritdoc />
        public void Send(string metric)
        {
            Send(new[] { metric });
        }

        /// <inheritdoc />
        public void Send(IEnumerable<string> metrics)
        {
            if (metrics == null)
            {
                return;
            }

            var data = EventArgsPool.Pop();

            if (data == null)
            {
                return;
            }

            var udpClient = _clientPool.Pop();

            if (udpClient == null)
            {
                return;
            }

            try
            {
                var endpoint = _endpointSource.GetEndpoint();

                bool connect = !udpClient.Client.Connected;

                // Is the endpoint unassigned or has changed?
                if (!endpoint.Equals(data.RemoteEndPoint))
                {
                    data.RemoteEndPoint = endpoint;
                    connect = true;
                }

                if (connect)
                {
                    udpClient.Client.Connect(data.RemoteEndPoint);
                }

                data.SendPacketsElements = metrics.ToMaximumBytePackets()
                    .Select(bytes => new SendPacketsElement(bytes, 0, bytes.Length, true))
                    .ToArray();

                udpClient.Client.SendPacketsAsync(data);
            }
            finally
            {
                _clientPool.Push(udpClient);
            }
        }

        private static UdpClient CreateUdpClient()
        {
            var client = new UdpClient();

            // See https://github.com/dotnet/corefx/pull/17853#issuecomment-291371266
            if (!RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
            {
                client.Client.SendBufferSize = 0;
            }

            return client;
        }
    }
}
```