@martincostello 2018-07-18
I'll have a quick go at porting it to `3.0.0` now and do a PR for review.