@AnthonySteele 2018-07-18
Concept proved, but the sockets are not thread-safe; abandoned in favour of pools.